Cater for PHPs HTTP_HOST possibly containing port numbers
Port numbers are appended to HTTP_HOST if they are non-standard.
They could theoretically be extracted from app.GetAllURLs() but I am not sure if it would be enough to just strip `http://`/ `https://` from these values or if they might also be followed by a path.
This takes a shortcut and does not whitelist the actual ports but rather allows ALL ports. I guess this is reasonable to keep the code simple.