Skip tools installation in chocolatey mode (Windows installer) (#1937)
* Skip tools installation in silent mode

Skips mkcert and nssm from installation in silent mode

* Force installation of sudo

This patch forces the installtion of sudo as DDEV can normaly not run correctly and full function without it. The license page is now showed before the component selection because it is not optional as before.

* Add Chocolatey mode

This patch adds a special Chocolatey mode which will prevent the installation of dependencies installed by Chocolatey.