Fix enable_xdebug / disable_xdebug not working for apache-cgi
Due to the nature of apache-cgi we need an config reload to make php
changes work. This adds an apache2 reload to the enable_xdebug and
disable_xdebug bash scripts in case the current webserver is apache-cgi.

This also updates the containertest.sh to check if both the CLI and the
webserver have Xdebug enabled/disabled after using the commands.