auth pantheon missing the dash
I was running through our setup documentation before on boarding a new dev. I was still running v1.10.2 and thought this may have changed in the current version, but after updating to v1.12.1 I was still running into an issue trying to configuration this using `ddev auth pantheon`. I'm not sure if this changed at some point and needs to be broken out by version, but `ddev auth-pantheon <YOUR TOKEN>` is required now.