Optimize bash handling
This patch optimizes the bash handling especially on Windows. Therefor the function FindWindowsBashPath() was replaced by FindBashPath() and a new function RunBashScriptInteractive() is added, both are now available in a separate implementation for Windows and Unix like systems so we can use the same code for each platform and do not have to make decisions during the runtime. To avoid later errors the bash binary provided by WSL is ignored. Also the check while adding the custom commands is reduced to the host service only and instead of directly returning continue with the next service.