Add support for persistent data on old Solr.
This provides support for persistent data on older versions of Solr. Tested with 5.

It seems like 5 stores data in `/opt/solr/server/solr/mycores/dev/data` while 8 stores data in `/var/solr/data/dev/data`.

This breaks the `The persistent volume should have the same name as the service so it can be deleted when the project is deleted.` part.

I'm not sure if `5` uses `/var/solr` at all, although there was stuff in it. Maybe an alternative is to add a comment to pick the relevant directory based on versions?