feat: add `file` option.
Motivation: While `export-db` uses `file` as the parameter to provide a name for the output file – the important part being that it eventually is indeed a FILE –, `import-db` uses `src`, even though it is also a parameter that eventually references a file (and not any other kind of source). For the sake of clarity and consistency these parameters should be nmed the same. Therefore `file` is added to `import-db` with the intention to substitute `src` completely at a later point.