Update pantheon.yaml.example
Adding debugging/verbose code that Randy walked me through when experiencing trying to do a ddev pull pantheon -y with a bad machine token. It wasn't clear that the token was bad, and it was returning a "Pull failed: exit status 1 " status that wasn't at all verbose/clear what was going on. The issue was a bad character in the machine token.