Revert "Add apt-get dist-upgrade -y to webimage_extra_packages (#3385)"
This exact dist-upgrade is now breaking dbimage_extra_packages.
It's probably due to an error in mariadb packages,
but it seems that apt-get dist-upgrade is not a risk-free
action.

Also, this whole problem seems to have been solved upstream in
the deb.sury.org package.

See https://github.com/drud/ddev/pull/3385 and
https://github.com/oerdnj/deb.sury.org/issues/1682

This reverts commit d385ef98feb6ddb98013cd484c55ab49e9094b3f.