Use bin/magento in magento command (#3699)
Currently the command looks for a script/binary "magento"
in the $PATH which is not available in a general Magento installation.
Change the script to use the shipped bin/magento.
If Magento is not installed via "composer install", the "bin/magento"
script is not available. So the check if ok with the right path.