refactor: Don’t defne `uploadDir` at all
Since `uploadDir` is a relative path that is relative to the docroot, the best we could do is return an empty string, meaning, it should be the same as the docroot, which is the default.

Returning a full path was wrong, because this should be a relative path. Also, returning “web” was wrong, because that’s already what the docroot is set to