Add more detail for Backdrop Quickstart
I was searching for the Backdrop Quickstart and it took me a while to find it buried inside the Drupal Quickstart. Then once I found it, it didn't have the information I was looking for (how to enable `bee`). Luckily I found that information elsewhere! 

Here's a PR to break Backdrop Quickstart out, and include the information about how to enable `bee`.