[tests only] Attempt to get actions/cache working again [skip ci]
There's a long term problem with actions/cache when
caching on Windows and then restoring on linux.
Apparently there is new trouble in this area.

But we were using the old actions/cache 3.0.11 on Windows,
and that suddenly seems to have stopped working.
They have struggled and struggled with this.