Add dev libraries / function test requirements
Quickstart needs to be updated if we want users to know how to run Functional / FunctionalJavascript tests.  We could do it this way or we could add this knowledge to the phpunit portion of the phpstorm documentation.  

Kind of feels like there should be a phpunit section instead of a phpunit in phpstorm section.