Tests and guidance for phpunit / Functional tests
Went through the process of getting phpunit + test debugging working for Functional tests and found that some additional documentation is needed to guide people through that process.  

This probably isn't the final documentation that will be committed but I want to get this in front of you as a starting point.