Some consistency changes to env variables in custom-commands.md
While adding DDEV_UID and DDEV_GID in a PR i've noticed that the env variable descriptions are sort of inconsistent. For one sometimes the first word after the colon is in lower case and sometimes in upper case. I thought after colon it should always be upper case? the only exception i've made for the description of DDEV_PROJECT_TYPE that just lists the cms versions which are names - same for DDEV_WEBSERVER_TYPE. And the other detail i wonder but wasn't sure in the end. would it make sense to end each description with a period cuz some already are? cuz the descriptions aren't always full sentences?