docs: Update Drupal 9 & 10 CMS Quickstarts
`ddev drush uli` generates a one-time link to the page for editing the uid1 (admin) account details. If the site is opened with that, there is no need to also run `ddev launch`, so that is confusing.

Hence, adding instructions on how to open the link generated by `ddev drush uli`, followed by a reference to `ddev launch` for further visits.