feat: streamline TYPO3 non-composer mode configuration support
TYPO3 v12 changed the naming of the configuration files
and also the place depending on the installation mode,
which differs between composer and non-composer setups.

Since #4268 ddev already uses the correct files and places
for composer based installations but not for non-composer
mode setups, which includes also TYPO3 Core Development
monorepo setups.

This change enhances the decetion for non-composer mode
installations to use correct setting files and paths to
mitigate the fact when no silent upgrades are executed
to migrate from the old to the new naming by TYPO3 [1]
and thus not containing the "releaxed" trustedPattern
required for ddev.

[1] https://review.typo3.org/c/Packages/TYPO3.CMS/+/81793