Make installing a Downloaded Binary in the $PATH clearer
If you follow the current instructions, you lose the path, since it is only temporary. So we should clarify that you need to actually  add an extra line in `~.bashrc.`

This is easy for us (more or less) but for people just starting with Unix and the command line, it's not easy at all. I remember it well.

So let's make it more welcoming for those not yet CLI veterans :)